is-significant-news/requirements.txt
pydantic
fastapi
requests
python-dotenv
